frontend/requirements.txt
reflex==0.8.17
python-dotenv==1.2.1